[876] Tighten up url boundaries
