+++ app/controllers/find/errors_controller.rb
@@ -1,0 +1,7 @@
+module Find
+  class ErrorsController < ApplicationController
+    include Errorable
+
+    layout "find_layout"
+  end
+end
