+++ app/controllers/publish/errors_controller.rb
@@ -1,0 +1,9 @@
+module Publish
+  class ErrorsController < ApplicationController
+    skip_before_action :authenticate
+
+    include Errorable
+
+    layout "publish"
+  end
+end
