+++ config/routes.rb
@@ -4,12 +4,6 @@ Rails.application.routes.draw do
   get :sha, controller: :heartbeat
   get :reporting, controller: :reporting
 
-  scope via: :all do
-    match "/404", to: "errors#not_found"
-    match "/500", to: "errors#internal_server_error"
-    match "/403", to: "errors#forbidden"
-  end
-
   constraints(APIConstraint.new) do
     get "/", to: redirect("/docs/")
   end
@@ -19,10 +13,14 @@ Rails.application.routes.draw do
   end
 
   if %w[development test review qa].include?(Rails.env)
-    draw(:find)
+    constraints(FindConstraint.new) do
+      draw(:find)
+    end
   end
 
-  draw(:publish)
-  draw(:support)
-  draw(:api)
+  constraints(PublishConstraint.new) do
+    draw(:publish)
+    draw(:support)
+    draw(:api)
+  end
 end
